Split events in their corresponding files
